Watt The Hack docs: hub + per-track subpages
Restructure /watt-the-hack/docs from the single Grid Simulator Mechanics page into a hub linking three track subpages: Base44 Pitching, City of Melbourne Grid Guardian (the relocated Grid Simulator Mechanics), and Amber Electric Smart Home. docs.tsx becomes a shell (backdrop + section switcher + Outlet); shared primitives extracted to app/components/watt-the-hack/docs/; the Docs nav becomes a group and /base44-pitching redirects to the docs subpage.
