+++ app/components/watt-the-hack/docs/primitives.tsx
@@ -1,0 +1,263 @@
+import type { ReactNode } from "react";
+
+export function SidebarLink({
+  href,
+  icon,
+  children,
+}: {
+  href: string;
+  icon: ReactNode;
+  children: ReactNode;
+}) {
+  return (
+    <a
+      href={href}
+      className="flex items-center gap-2.5 rounded-md px-2 py-1.5 text-[13px] font-medium text-muted transition-colors hover:bg-subtle hover:text-ink"
+    >
+      <span className="text-sky-500">{icon}</span>
+      <span className="min-w-0 truncate">{children}</span>
+    </a>
+  );
+}
+
+export function ComponentCard({
+  title,
+  icon,
+  bgColor,
+  children,
+}: {
+  title: string;
+  icon: ReactNode;
+  bgColor: string;
+  children: ReactNode;
+}) {
+  return (
+    <div className="rounded-xl border border-line bg-surface p-4 shadow-sm transition-shadow hover:shadow-md">
+      <div className="flex items-center gap-3">
+        <div className={`flex h-9 w-9 items-center justify-center rounded-lg ${bgColor}`}>{icon}</div>
+        <h3 className="text-sm font-semibold text-ink">{title}</h3>
+      </div>
+      <div className="mt-2.5 text-[13px] leading-relaxed text-muted">{children}</div>
+    </div>
+  );
+}
+
+export function ConstraintRow({
+  title,
+  children,
+  tone = "default",
+}: {
+  title: string;
+  children: ReactNode;
+  tone?: "default" | "warning" | "danger";
+}) {
+  const tones = {
+    default: {
+      card: "border-sky-300 bg-sky-100",
+      title: "text-sky-950",
+      body: "text-slate-700",
+      accent: "bg-sky-400",
+    },
+    warning: {
+      card: "border-amber-300 bg-amber-100",
+      title: "text-amber-950",
+      body: "text-amber-950",
+      accent: "bg-amber-400",
+    },
+    danger: {
+      card: "border-rose-300 bg-rose-100",
+      title: "text-rose-950",
+      body: "text-rose-950",
+      accent: "bg-rose-400",
+    },
+  };
+  const style = tones[tone];
+
+  return (
+    <div className={`relative flex flex-col gap-2 overflow-hidden rounded-lg border pl-5 pr-4 py-3 sm:flex-row sm:items-start sm:gap-4 ${style.card}`}>
+      <div className={`absolute inset-y-0 left-0 w-1.5 ${style.accent}`} />
+      <div className="sm:w-1/3">
+        <h4 className={`font-semibold ${style.title}`}>{title}</h4>
+      </div>
+      <div className={`min-w-0 text-sm leading-relaxed sm:w-2/3 ${style.body}`}>{children}</div>
+    </div>
+  );
+}
+
+const COST_REFERENCE: {
+  group: string;
+  rows: { key: string; what: string; rate: string; income?: boolean }[];
+}[] = [
+  {
+    group: "Core — charged in every scenario",
+    rows: [
+      {
+        key: "tariff_import",
+        what: "Buying power from the external grid.",
+        rate: 'price × MWh imported — dynamic, read state["price"] ($/MWh)',
+      },
+      {
+        key: "tariff_export",
+        what: "Selling surplus solar or battery power back to the grid.",
+        rate: "$50 / MWh exported",
+        income: true,
+      },
+      {
+        key: "demand_charge",
+        what: "Your single biggest import spike, billed once for the whole run (peak-shaving discipline).",
+        rate: "$1,000 / MW of peak import",
+      },
+      {
+        key: "carbon_cost",
+        what: "Carbon price on CO₂ from grid imports + diesel. Scenarios can override grid intensity.",
+        rate: "$50 / kg CO₂ · grid ≈ 0.7, diesel ≈ 0.27 kg/MWh",
+      },
+      {
+        key: "battery_wear",
+        what: "Wear from cycling the battery — charging or discharging.",
+        rate: "$50 / MWh moved",
+      },
+      {
+        key: "ramp_charge",
+        what: "Smoothness penalty on the step-to-step change in net grid power.",
+        rate: "(ΔMW)² × $1",
+      },
+      {
+        key: "generator_fuel",
+        what: "Running the diesel generator.",
+        rate: "$1,000 / MWh",
+      },
+      {
+        key: "blackout_penalty",
+        what: "Unmet demand (load shed). Avoid at all costs.",
+        rate: "$100,000 / MWh",
+      },
+      {
+        key: "overvoltage_penalty",
+        what: "Exporting beyond the grid export cap.",
+        rate: "$5,000 / MWh",
+      },
+    ],
+  },
+  {
+    group: "FCAS — when the scenario enables the reserve market",
+    rows: [
+      {
+        key: "fcas_revenue",
+        what: "Availability pay for capacity you keep on standby (whether or not it's called).",
+        rate: "$40 / MW / hour reserved",
+        income: true,
+      },
+      {
+        key: "fcas_dispatch_bonus",
+        what: "Energy you actually deliver when a dispatch event calls your reserve.",
+        rate: "$200 / MWh delivered",
+        income: true,
+      },
+      {
+        key: "fcas_shortfall_penalty",
+        what: "Failing to deliver a called dispatch — bid only what your SOC can back.",
+        rate: "$100,000 / MWh short",
+      },
+      {
+        key: "fcas_ramp_charge",
+        what: "Volatility in your reserve bid between steps.",
+        rate: "$500 / MW of change",
+      },
+    ],
+  },
+  {
+    group: "Advanced scenarios only — $0 unless that mechanic is live",
+    rows: [
+      {
+        key: "compliance_penalty",
+        what: "Breaching an Operator's-Mandate window you opted into (SOC floor / export cap).",
+        rate: "$2,000,000 / SOC-unit · $500,000 / MW over cap, per step",
+      },
+      {
+        key: "diesel_ban_penalty",
+        what: "Running diesel during a ban with no valid agent_plan exemption.",
+        rate: "$3,000 / MWh",
+      },
+      {
+        key: "anomaly_ack_fine",
+        what: "Each step inside an anomaly window you didn't acknowledge in agent_plan.",
+        rate: "$5,000 / step",
+      },
+      {
+        key: "cyber_containment_fine",
+        what: "Missing a real attack, or acknowledging a fake one.",
+        rate: "$50,000 each",
+      },
+      {
+        key: "ids_cost",
+        what: "Subscribing to the intrusion-detection signal.",
+        rate: "flat per-step fee set by the scenario",
+      },
+      {
+        key: "phishing_fine",
+        what: "Acting on a phishing / bait directive.",
+        rate: "fine set by the scenario",
+      },
+    ],
+  },
+];
+
+export function CostReferenceTable() {
+  return (
+    <div className="rounded-lg border border-slate-300 bg-white px-5 py-4">
+      <h4 className="font-semibold text-slate-900">Cost &amp; penalty reference</h4>
+      <p className="mt-1 text-sm leading-relaxed text-slate-700">
+        Every line that can appear in your <code>cost_breakdown</code>, with the exact rate the engine applies.
+        Positive numbers are costs; <span className="font-semibold text-emerald-700">green is income</span>. Most rows
+        read $0 unless their mechanic is active in the scenario you&apos;re running — the bottom two groups only fire in
+        the advanced scenarios.
+      </p>
+      <div className="mt-3 overflow-x-auto">
+        <table className="w-full border-collapse text-left text-[13px]">
+          <thead>
+            <tr className="border-b border-slate-300 text-[11px] uppercase tracking-wide text-slate-500">
+              <th className="py-2 pr-4 font-semibold">Line item</th>
+              <th className="py-2 pr-4 font-semibold">What it is</th>
+              <th className="py-2 font-semibold">Rate</th>
+            </tr>
+          </thead>
+          {COST_REFERENCE.map((section) => (
+            <tbody key={section.group}>
+              <tr>
+                <td
+                  colSpan={3}
+                  className="pt-4 pb-1 text-[11px] font-bold uppercase tracking-wider text-sky-700"
+                >
+                  {section.group}
+                </td>
+              </tr>
+              {section.rows.map((row) => (
+                <tr key={row.key} className="border-t border-slate-200 align-top">
+                  <td className="py-2 pr-4">
+                    <code>{row.key}</code>
+                  </td>
+                  <td className="py-2 pr-4 text-slate-700">{row.what}</td>
+                  <td
+                    className={`py-2 font-medium tabular-nums ${row.income ? "text-emerald-700" : "text-slate-900"}`}
+                  >
+                    {row.rate}
+                    {row.income ? " (income)" : ""}
+                  </td>
+                </tr>
+              ))}
+            </tbody>
+          ))}
+        </table>
+      </div>
+    </div>
+  );
+}
+
+export function CodeBlock({ children }: { children: ReactNode }) {
+  return (
+    <pre className="overflow-x-auto rounded-lg bg-[#0E1525] p-3.5 font-mono text-[12.5px] leading-relaxed text-slate-300">
+      <code>{children}</code>
+    </pre>
+  );
+}
